amd64: Change "pointer" fields in java objects from int to long.

+++ jbinding-java/src/net/sf/sevenzipjbinding/impl/InArchiveImpl.java
@@ -80,10 +80,10 @@ public class InArchiveImpl implements ISevenZipInArchive {
 	}
 
 	@SuppressWarnings("unused")
-	private int sevenZipArchiveInstance;
+	private long sevenZipArchiveInstance;
 
 	@SuppressWarnings("unused")
-	private int sevenZipArchiveInStreamInstance;
+	private long sevenZipArchiveInStreamInstance;
 
 	private int numberOfItems = -1;
